+++ doc/changelog.md
@@ -7,6 +7,7 @@
 - Support for additional card list database websites
   - MTGDecks (mtgdecks.net)
   - Archidekt (archidekt.com)
+  - TCGPlayer Infinite (infinite.tcgplayer.com)
 
 # Version 0.21.0 (2023-02-08)  <a name="v0_21_0"></a>